+++ packages/common/src/utils/domUtils.ts
@@ -176,7 +176,7 @@ export function setControlValue(
     }
   }
 
-  el.value = String(fieldValue || '');
+  el.value = String(fieldValue ?? '');
 }
 
 /** Sets the form inputs value to match the data object provided. */
@@ -192,7 +192,7 @@ export function setForm<Data extends Obj>(
   }
 }
 
-type ErrorField = string | Record<string, unknown> | string[];
+type ErrorField = string | Obj | string[];
 
 function executeCustomizer(objValue?: ErrorField, srcValue?: ErrorField) {
   if (_isPlainObject(objValue) || _isPlainObject(srcValue)) return;