+++ packages/site/src/routes/__layout.svelte
@@ -1,4 +1,6 @@
 <script>
+  import 'focus-visible';
+  import '../app.postcss';
   import Nav from '$lib/components/Nav.svelte';
   import 'tippy.js/dist/tippy.css';
   import 'highlight.js/styles/a11y-dark.css';