+++ packages/core/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @felte/core
 
+## 0.1.6
+
+### Patch Changes
+
+- 5bb4a02: Add data-felte-ignore attribute to make Felte completely ignore an input
+- Updated dependencies [5bb4a02]
+  - @felte/common@0.4.10
+
+## 0.1.5
+
+### Patch Changes
+
+- 6b8aafb: Fix `validate` function on `createSubmitHandler` config not updating errors on store
+
+## 0.1.4
+
+### Patch Changes
+
+- 16ff018: Export ES module as default
+- Updated dependencies [16ff018]
+  - @felte/common@0.4.9
+
+## 0.1.3
+
+### Patch Changes
+
+- e6034c0: Use `Record<string, any>` for config type in order to allow interfaces to be passed as types
+- Updated dependencies [af4b183]
+- Updated dependencies [e6034c0]
+  - @felte/common@0.4.8
+
 ## 0.1.2
 
 ### Patch Changes