+++ packages/reporter-solid/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @felte/reporter-solid
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [5bb4a02]
+  - @felte/common@0.4.10
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [16ff018]
+  - @felte/common@0.4.9
+
+## 0.1.6
+
+### Patch Changes
+
+- 14bf9d8: Update to use only JS features compatible with Bundlephobia
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [af4b183]
+- Updated dependencies [e6034c0]
+  - @felte/common@0.4.8
+
 ## 0.1.4
 
 ### Patch Changes