+++ packages/reporter-svelte/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @felte/reporter-dom
 
+## 0.3.14
+
+### Patch Changes
+
+- Updated dependencies [5bb4a02]
+  - @felte/common@0.4.10
+
+## 0.3.13
+
+### Patch Changes
+
+- 16ff018: Export ES module as default
+- Updated dependencies [16ff018]
+  - @felte/common@0.4.9
+
+## 0.3.12
+
+### Patch Changes
+
+- 14bf9d8: Update to use only JS features compatible with Bundlephobia
+
+## 0.3.11
+
+### Patch Changes
+
+- Updated dependencies [af4b183]
+- Updated dependencies [e6034c0]
+  - @felte/common@0.4.8
+
 ## 0.3.10
 
 ### Patch Changes