+++ packages/validator-yup/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @felte/validator-yup
 
+## 0.2.7
+
+### Patch Changes
+
+- 0bce41e: Fix rollup config for validators
+
+## 0.2.6
+
+### Patch Changes
+
+- 16ff018: Export ES module as default
+- Updated dependencies [16ff018]
+  - @felte/common@0.4.9
+
 ## 0.2.5
 
 ### Patch Changes